Remove the MultiLevelMenu class - it's doing way too much, and sometimes unnecessarily. Instead, add some mixins for the defining similar behaviour, and use where appropriate
